backend/pdftoexcel/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
PyPDF2>=3.0.0
tabula-py>=2.9.0
pandas>=2.0.0
openpyxl>=3.1.0
